[DOM-280] Add HighlightableScaleView
